[iris] Point importers at the renamed transition_reader module
The rename commit landed the file move (worker_source.py -> transition_reader.py)
but a botched git add dropped the matching import updates, so backend_store.py,
composer.py, and the cluster conftest still imported the now-missing
iris.cluster.controller.worker_source. Update them to transition_reader.
