+++ lib/iris/src/iris/cluster/composer.py
@@ -36,7 +36,7 @@ from iris.cluster.controller.backend import TaskBackend
 from iris.cluster.controller.db import ControllerDB
 from iris.cluster.controller.log_stack import LogStack
 from iris.cluster.controller.reconcile.loader import TransitionReader
-from iris.cluster.controller.worker_source import DbTransitionReader
+from iris.cluster.controller.transition_reader import DbTransitionReader
 from iris.cluster.inject_env import TASK_ENV_SECRET_NAME, projects_task_env_secret
 from iris.cluster.platforms.factory import ProviderBundle, create_provider_bundle
 from iris.cluster.platforms.k8s.service import CloudK8sService
